refactor: split oversized call and type printer shards
