+++ crates/tsz-checker/src/types/computation/call/inner_argument_collection.rs
@@ -1,0 +1,1587 @@
+{
+        let mut non_generic_contextual_types: Option<Vec<Option<TypeId>>> = None;
+        // Track whether we pushed a ThisType marker to this_type_stack during call processing.
+        let mut pushed_this_type_from_shape = false;
+        // Extracted ThisType<T> from shape params (via alias expansion if needed).
+        // Saved so it can be re-pushed around retry arg typing calls.
+        let mut shape_this_type: Option<TypeId> = None;
+        // Track whether Round 2 successfully used a non-empty return context substitution.
+        // When true, the post-inference retry should be suppressed because Round 2 already
+        // correctly resolved the callback parameter types using the return context.
+        let mut had_return_context_substitution = false;
+        // Checker-side intra-expression Round 2 substitution. When the two-pass logic
+        // builds a substitution that the solver's single-pass `resolve_call` cannot
+        // recover (e.g., a homomorphic mapped+infer return position drops the
+        // setup-derived inference), we use this to refine the solver's
+        // `instantiated_params` so the post-call assignability recheck sees the
+        // tighter expected type. See `intra_expression_inference_homomorphic_mapped_return_type`.
+        let mut checker_round2_substitution: Option<
+            crate::query_boundaries::common::TypeSubstitution,
+        > = None;
+        // Snapshot of the original shape (pre-solver), used to re-instantiate params
+        // when overriding the solver with `checker_round2_substitution`.
+        let mut checker_round2_shape: Option<common::FunctionShape> = None;
+        let mut direct_literal_conflict_substitution: Option<common::TypeSubstitution> = None;
+        let arg_types = if is_generic_call {
+            if let Some(shape) = callee_shape {
+                // Pre-compute which parameter positions should skip excess property
+                // checking because the original parameter type contains a type parameter.
+                // For generic calls like `parrot<T extends Named>({name, sayHello(){}})`,
+                // the instantiated type is the constraint `Named`, but tsc skips excess
+                // property checks because `T` captures the full object type.
+                //
+                // Use the raw FunctionShape parameter types (which preserve type parameters)
+                // rather than ctx_helper.get_parameter_type_for_call (which may resolve
+                // through Lazy/Application types and lose type parameter information).
+                let needs_generic_skip_state = args.iter().copied().any(|arg_idx| {
+                    self.ctx
+                        .arena
+                        .get(self.ctx.arena.skip_parenthesized_and_assertions(arg_idx))
+                        .is_some_and(|node| {
+                            node.kind == syntax_kind_ext::OBJECT_LITERAL_EXPRESSION
+                                || node.kind == SyntaxKind::StringLiteral as u16
+                                || node.kind == SyntaxKind::NumericLiteral as u16
+                                || node.kind == SyntaxKind::BigIntLiteral as u16
+                                || node.kind == SyntaxKind::TrueKeyword as u16
+                                || node.kind == SyntaxKind::FalseKeyword as u16
+                                || node.kind == SyntaxKind::NoSubstitutionTemplateLiteral as u16
+                        })
+                });
+                if needs_generic_skip_state {
+                    let arg_count = args.len();
+                    let excess_skip: Vec<bool> = (0..arg_count)
+                        .map(|i| {
+                            // Check if the parameter type IS a bare type parameter (T),
+                            // not just a complex type that contains type parameters
+                            // (like a mapped type). Only bare type parameters capture
+                            // the full object shape, making excess property checking
+                            // inappropriate.
+                            let from_shape = if i < shape.params.len() {
+                                is_type_parameter_type(self.ctx.types, shape.params[i].type_id)
+                            } else if let Some(last) = shape.params.last() {
+                                // Rest parameter: check the rest param's type
+                                last.rest && is_type_parameter_type(self.ctx.types, last.type_id)
+                            } else {
+                                false
+                            };
+                            let from_ctx = ctx_helper
+                                .get_parameter_type_for_call(i, arg_count)
+                                .is_some_and(|param_type| {
+                                    is_type_parameter_type(self.ctx.types, param_type)
+                                });
+                            from_shape || from_ctx
+                        })
+                        .collect();
+                    let has_any_excess_skip = excess_skip.iter().any(|&s| s);
+                    if has_any_excess_skip {
+                        self.ctx.generic_excess_skip = Some(excess_skip);
+                    }
+                }
+
+                // Pre-compute which arguments are contextually sensitive to avoid borrowing self in closures.
+                let sensitive_args: Vec<bool> = args
+                    .iter()
+                    .map(|&arg| is_contextually_sensitive(self, arg))
+                    .collect();
+                let suppress_generic_return_context = contextual_type.is_some()
+                    && (args
+                        .iter()
+                        .copied()
+                        .any(|arg| self.suppress_generic_return_context_for_arg(arg))
+                        || self.suppress_generic_return_context_for_direct_arg_overlap(
+                            &shape,
+                            args,
+                            contextual_type,
+                        ));
+                let generic_inference_contextual_type = if suppress_generic_return_context {
+                    None
+                } else {
+                    // TypeScript ignores `any` and `unknown` contextual return types
+                    // during generic inference — they carry no useful information and
+                    // can cause type parameters to be inferred as `any`/`unknown`,
+                    // which destroys contextual typing for callback parameters
+                    // (e.g., `T & ((arg: string) => any)` becoming just `any`).
+                    contextual_type.filter(|&ct| ct != TypeId::ANY && ct != TypeId::UNKNOWN)
+                };
+                trace!(
+                    type_params = ?shape
+                        .type_params
+                        .iter()
+                        .map(|tp| self.ctx.types.resolve_atom(tp.name))
+                        .collect::<Vec<_>>(),
+                    generic_inference_contextual_type = ?generic_inference_contextual_type.map(|t| t.0),
+                    suppress_generic_return_context,
+                    "Generic call contextual type gate"
+                );
+                let round1_skip_outer_context: Vec<bool> = args
+                    .iter()
+                    .map(|&arg| self.round1_should_skip_outer_contextual_type(arg))
+                    .collect();
+                let needs_two_pass = sensitive_args.iter().copied().any(std::convert::identity);
+
+                // Extract ThisType<T> from shape params via alias expansion.
+                // Store for re-use across retry arg typing calls.
+                if shape_this_type.is_none() {
+                    let env = self.ctx.type_env.borrow();
+                    for param in &shape.params {
+                        let ctx_helper = ContextualTypeContext::with_expected_and_options(
+                            self.ctx.types,
+                            param.type_id,
+                            self.ctx.compiler_options.no_implicit_any,
+                        );
+                        // Try simple extraction first, then with alias expansion
+                        // (e.g., ConstructorOptions<Data> → ... & ThisType<Instance<Data>>).
+                        let tt = ctx_helper
+                            .get_this_type_from_marker()
+                            .or_else(|| ctx_helper.get_this_type_from_marker_with_resolver(&*env));
+                        if let Some(tt) = tt {
+                            shape_this_type = Some(tt);
+                            break;
+                        }
+                    }
+                }
+
+                if needs_two_pass {
+                    // Round 1: skip sensitive args so inference can start from concrete inputs.
+                    let round1_diag_snap = contextual_type
+                        .is_some()
+                        .then(|| self.ctx.snapshot_diagnostics());
+                    // Also snapshot TS2454 dedup state so that rolling back Round 1
+                    // diagnostics allows TS2454 to be re-emitted during Round 2.
+                    // Without this, TS2454 emitted during Round 1 gets removed from
+                    // the diagnostics vector but remains in the dedup set, silencing
+                    // the error in Round 2.
+                    let round1_ts2454_snap = round1_diag_snap
+                        .as_ref()
+                        .map(|_| self.ctx.emitted_ts2454_errors.clone());
+                    let mut round1_arg_types = self.collect_call_argument_types_with_context(
+                        args,
+                        |i, _arg_count| {
+                            let skip_round1_context = (i < sensitive_args.len()
+                                && sensitive_args[i])
+                                || (i < round1_skip_outer_context.len()
+                                    && round1_skip_outer_context[i]);
+                            if skip_round1_context {
+                                None
+                            } else {
+                                base_contextual_param_types.get(i).copied().flatten()
+                            }
+                        },
+                        check_excess_properties,
+                        Some(&sensitive_args), // Skip sensitive args in Round 1
+                        callable_ctx,
+                    );
+
+                    // Pre-compute type parameter names and set once for reuse
+                    // across Round 1/Round 2 inference passes.
+                    let type_param_names: Vec<tsz_common::Atom> =
+                        shape.type_params.iter().map(|tp| tp.name).collect();
+                    let tracked_type_params: FxHashSet<tsz_common::Atom> =
+                        type_param_names.iter().copied().collect();
+
+                    let direct_literal_conflict_type_params = self
+                        .direct_round1_literal_preservation_type_params(
+                            &shape,
+                            args,
+                            &round1_arg_types,
+                            &sensitive_args,
+                        );
+                    if !direct_literal_conflict_type_params.is_empty() {
+                        direct_literal_conflict_substitution =
+                            Some(direct_literal_conflict_type_params.clone());
+                        checker_round2_shape = Some(shape.clone());
+                    }
+
+                    // Seed inference from non-sensitive object-literal properties.
+                    let mut extracted_round1_partials = vec![false; args.len()];
+                    for (i, &arg_idx) in args.iter().enumerate() {
+                        if !sensitive_args[i] {
+                            continue;
+                        }
+                        let param_type = shape.params.get(i).map(|p| p.type_id).or_else(|| {
+                            let last = shape.params.last()?;
+                            last.rest.then_some(last.type_id)
+                        });
+                        let partial = param_type
+                            .and_then(|param_type| {
+                                self.extract_inference_contributing_object_type(
+                                    arg_idx,
+                                    param_type,
+                                    &type_param_names,
+                                )
+                                .or_else(|| {
+                                    self.extract_inference_contributing_array_type(
+                                        arg_idx,
+                                        param_type,
+                                        &type_param_names,
+                                    )
+                                })
+                            })
+                            .or_else(|| self.extract_non_sensitive_object_type(arg_idx));
+                        if let Some(partial) = partial {
+                            round1_arg_types[i] = partial;
+                            extracted_round1_partials[i] = true;
+                        }
+                    }
+                    // If only callback members remain, include ones with concrete param context.
+                    for (i, &arg_idx) in args.iter().enumerate() {
+                        if !sensitive_args[i] || extracted_round1_partials[i] {
+                            continue;
+                        }
+                        let Some(param_type) =
+                            shape.params.get(i).map(|p| p.type_id).or_else(|| {
+                                let last = shape.params.last()?;
+                                last.rest.then_some(last.type_id)
+                            })
+                        else {
+                            continue;
+                        };
+                        if let Some(partial) = self
+                            .extract_inference_contributing_object_type(
+                                arg_idx,
+                                param_type,
+                                &type_param_names,
+                            )
+                            .or_else(|| {
+                                self.extract_inference_contributing_array_type(
+                                    arg_idx,
+                                    param_type,
+                                    &type_param_names,
+                                )
+                            })
+                        {
+                            round1_arg_types[i] = partial;
+                            extracted_round1_partials[i] = true;
+                        }
+                    }
+                    for (i, arg_type) in round1_arg_types.iter_mut().enumerate() {
+                        if !sensitive_args.get(i).copied().unwrap_or(false) {
+                            continue;
+                        }
+                        let Some(arg_node) = self.ctx.arena.get(args[i]) else {
+                            continue;
+                        };
+                        if !self.is_callback_like_argument(args[i]) {
+                            continue;
+                        }
+                        if self
+                            .ctx
+                            .arena
+                            .get_function(arg_node)
+                            .and_then(|func| func.type_parameters.as_ref())
+                            .is_some_and(|params| !params.nodes.is_empty())
+                        {
+                            let raw_arg_type =
+                                self.get_type_of_node_with_request(args[i], &TypingRequest::NONE);
+                            let seeded =
+                                self.sanitize_generic_inference_arg_type(args[i], raw_arg_type);
+                            if seeded != TypeId::UNKNOWN && seeded != TypeId::ERROR {
+                                *arg_type = seeded;
+                                continue;
+                            }
+                        }
+                        let Some(param_type) =
+                            shape.params.get(i).map(|p| p.type_id).or_else(|| {
+                                let last = shape.params.last()?;
+                                last.rest.then_some(last.type_id)
+                            })
+                        else {
+                            continue;
+                        };
+                        if extracted_round1_partials.get(i).copied().unwrap_or(false) {
+                            continue;
+                        }
+                        if self.sensitive_callback_placeholder_should_skip_round1_inference(
+                            &shape, param_type,
+                        ) {
+                            *arg_type = TypeId::UNKNOWN;
+                        }
+                    }
+                    // Nested calls whose outer contextual type was intentionally skipped in
+                    // Round 1 should not poison outer inference with provisional `error` or
+                    // `__infer_*` results. Leave them for Round 2 unless they resolved cleanly.
+                    for (i, arg_type) in round1_arg_types.iter_mut().enumerate() {
+                        if !round1_skip_outer_context.get(i).copied().unwrap_or(false) {
+                            continue;
+                        }
+                        let unresolved =
+                            common::is_unresolved_inference_result(self.ctx.types, *arg_type);
+                        if unresolved {
+                            *arg_type = TypeId::UNKNOWN;
+                        }
+                    }
+                    // === Perform Round 1 Inference ===
+                    // Pre-evaluate function shape parameter types through the
+                    // TypeEnvironment so the solver can constrain against concrete
+                    // object types instead of unresolved Application types.
+                    // Example: Opts<State, Actions> → { state?: State, actions: Actions }
+                    let evaluated_shape = {
+                        let new_params: Vec<_> = shape
+                            .params
+                            .iter()
+                            .enumerate()
+                            .map(|(i, p)| {
+                                let arg_type = round1_arg_types.get(i).copied();
+                                let preserve_raw_application = arg_type.is_some_and(|arg_type| {
+                                    common::should_preserve_application_for_inference(
+                                        self.ctx.types,
+                                        p.type_id,
+                                        arg_type,
+                                    )
+                                });
+
+                                common::ParamInfo {
+                                    name: p.name,
+                                    type_id: if preserve_raw_application {
+                                        p.type_id
+                                    } else {
+                                        self.evaluate_type_with_env(p.type_id)
+                                    },
+                                    optional: p.optional,
+                                    rest: p.rest,
+                                }
+                            })
+                            .collect();
+                        common::FunctionShape {
+                            params: new_params,
+                            return_type: shape.return_type,
+                            this_type: shape.this_type,
+                            type_params: shape.type_params.clone(),
+                            type_predicate: shape.type_predicate,
+                            is_constructor: shape.is_constructor,
+                            is_method: shape.is_method,
+                        }
+                    };
+                    let mut substitution = {
+                        let env = self.ctx.type_env.borrow();
+                        call_checker::compute_contextual_types_with_context(
+                            self.ctx.types,
+                            &self.ctx,
+                            &env,
+                            &evaluated_shape,
+                            &round1_arg_types,
+                            generic_inference_contextual_type,
+                        )
+                    };
+
+                    // Partial object literals can expose inference from callback
+                    // returns before the full Round 2 contextual pass.
+                    for (i, &arg_type) in round1_arg_types.iter().enumerate() {
+                        if !extracted_round1_partials.get(i).copied().unwrap_or(false)
+                            || arg_type == TypeId::UNKNOWN
+                            || arg_type == TypeId::ERROR
+                        {
+                            continue;
+                        }
+                        let Some(param_type) = evaluated_shape.params.get(i).map(|p| p.type_id)
+                        else {
+                            continue;
+                        };
+                        let mut partial_substitution =
+                            crate::query_boundaries::common::TypeSubstitution::new();
+                        let mut visited = FxHashSet::default();
+                        self.collect_return_context_substitution(
+                            param_type,
+                            arg_type,
+                            &tracked_type_params,
+                            &mut partial_substitution,
+                            &mut visited,
+                        );
+                        for (&name, &ty) in partial_substitution.map() {
+                            if ty == TypeId::UNKNOWN
+                                || ty == TypeId::ERROR
+                                || self.target_contains_blocking_return_context_type_params(
+                                    ty,
+                                    &tracked_type_params,
+                                )
+                            {
+                                continue;
+                            }
+                            if substitution.get(name).is_none_or(|existing| {
+                                existing == TypeId::UNKNOWN
+                                    || existing == TypeId::ERROR
+                                    || common::contains_type_parameters(self.ctx.types, existing)
+                                    || common::contains_infer_types(self.ctx.types, existing)
+                            }) {
+                                substitution.insert(name, ty);
+                            }
+                        }
+                    }
+
+                    // Extract ThisType<T> marker from raw parameter types and
+                    // instantiate with the Round 1 substitution. Push to
+                    // this_type_stack so nested object literal methods resolve
+                    // `this` to the inferred type.
+                    if !pushed_this_type_from_shape {
+                        if let Some(tt) = shape_this_type {
+                            let instantiated = crate::query_boundaries::common::instantiate_type(
+                                self.ctx.types,
+                                tt,
+                                &substitution,
+                            );
+                            self.ctx.this_type_stack.push(instantiated);
+                            pushed_this_type_from_shape = true;
+                        } else {
+                            let env = self.ctx.type_env.borrow();
+                            for param in &shape.params {
+                                let ctx_helper = ContextualTypeContext::with_expected_and_options(
+                                    self.ctx.types,
+                                    param.type_id,
+                                    self.ctx.compiler_options.no_implicit_any,
+                                );
+                                let this_type =
+                                    ctx_helper.get_this_type_from_marker().or_else(|| {
+                                        ctx_helper.get_this_type_from_marker_with_resolver(&*env)
+                                    });
+                                if let Some(this_type) = this_type {
+                                    let instantiated =
+                                        crate::query_boundaries::common::instantiate_type(
+                                            self.ctx.types,
+                                            this_type,
+                                            &substitution,
+                                        );
+                                    self.ctx.this_type_stack.push(instantiated);
+                                    pushed_this_type_from_shape = true;
+                                    break;
+                                }
+                            }
+                            drop(env);
+                        }
+                    }
+
+                    for (i, &arg_idx) in args.iter().enumerate() {
+                        if !sensitive_args.get(i).copied().unwrap_or(false) {
+                            continue;
+                        }
+                        let Some(_) = self.ctx.arena.get(arg_idx) else {
+                            continue;
+                        };
+                        if !self.is_callback_like_argument(arg_idx) {
+                            continue;
+                        }
+                        let Some(param_type) =
+                            shape.params.get(i).map(|p| p.type_id).or_else(|| {
+                                let last = shape.params.last()?;
+                                last.rest.then_some(last.type_id)
+                            })
+                        else {
+                            continue;
+                        };
+
+                        let names_to_strip: Vec<_> = shape
+                            .type_params
+                            .iter()
+                            .filter_map(|tp| {
+                                substitution.get(tp.name).and_then(|inferred| {
+                                    self.should_strip_sensitive_placeholder_substitution(
+                                        &shape, param_type, tp.name, inferred,
+                                    )
+                                    .then_some(tp.name)
+                                })
+                            })
+                            .collect();
+                        if !names_to_strip.is_empty() {
+                            let names_to_strip: rustc_hash::FxHashSet<_> =
+                                names_to_strip.into_iter().collect();
+                            let mut filtered =
+                                crate::query_boundaries::common::TypeSubstitution::new();
+                            for (&name, &type_id) in substitution.map() {
+                                if !names_to_strip.contains(&name) {
+                                    filtered.insert(name, type_id);
+                                }
+                            }
+                            substitution = filtered;
+                        }
+                    }
+                    let inferred_type_params_by_name: Vec<_> = shape
+                        .type_params
+                        .iter()
+                        .filter_map(|tp| {
+                            substitution
+                                .get(tp.name)
+                                .map(|ty| (self.ctx.types.resolve_atom(tp.name), ty))
+                        })
+                        .collect();
+                    trace!(
+                        substitution_is_empty = substitution.is_empty(),
+                        "Round 1 inference: substitution computed"
+                    );
+                    let mut round2_substitution = substitution.clone();
+                    let mut round2_return_context_names = FxHashSet::default();
+                    if let Some(ctx_type) = generic_inference_contextual_type {
+                        let mut return_context_substitution = self
+                            .compute_return_context_substitution_from_shape(&shape, Some(ctx_type));
+                        let return_param_names: FxHashSet<_> = self
+                            .function_like_return_parameter_type_params(&shape)
+                            .into_iter()
+                            .collect();
+                        let same_return_context_application =
+                            common::application_info(self.ctx.types, shape.return_type)
+                                .zip(common::application_info(self.ctx.types, ctx_type))
+                                .is_some_and(|((return_base, _), (ctx_base, _))| {
+                                    return_base == ctx_base
+                                });
+                        let return_context_specializes_return_params = !return_param_names
+                            .is_empty()
+                            && self.contextual_return_type_specializes_wrapped_params(
+                                shape.return_type,
+                                ctx_type,
+                                &return_param_names,
+                                &mut FxHashSet::default(),
+                            );
+                        if !return_param_names.is_empty()
+                            && !same_return_context_application
+                            && !return_context_specializes_return_params
+                        {
+                            let mut filtered =
+                                crate::query_boundaries::common::TypeSubstitution::new();
+                            for (&name, &type_id) in return_context_substitution.map() {
+                                if !return_param_names.contains(&name) {
+                                    filtered.insert(name, type_id);
+                                }
+                            }
+                            return_context_substitution = filtered;
+                        }
+                        trace!(
+                            type_params = ?shape
+                                .type_params
+                                .iter()
+                                .map(|tp| self.ctx.types.resolve_atom(tp.name))
+                                .collect::<Vec<_>>(),
+                            contextual_type = ctx_type.0,
+                            contextual_type_display = %self.format_type(ctx_type),
+                            contextual_type_union_members = ?common::union_members(
+                                self.ctx.types,
+                                ctx_type,
+                            )
+                            .map(|members| members
+                                .into_iter()
+                                .map(|member| (
+                                    self.format_type(member),
+                                    query::get_application_info(self.ctx.types, member)
+                                        .map(|(_, args)| args),
+                                ))
+                                .collect::<Vec<_>>()),
+                            return_type_display = %self.format_type(shape.return_type),
+                            return_context_substitution = ?return_context_substitution
+                                .map()
+                                .iter()
+                                .map(|(name, ty)| (self.ctx.types.resolve_atom(*name), ty.0))
+                                .collect::<Vec<_>>(),
+                            "Round 2 return-context substitution"
+                        );
+                        for (&name, &ty) in return_context_substitution.map().iter() {
+                            if ty == TypeId::UNKNOWN
+                                || ty == TypeId::ERROR
+                                || self.target_contains_blocking_return_context_type_params(
+                                    ty,
+                                    &tracked_type_params,
+                                )
+                            {
+                                continue;
+                            }
+
+                            let should_update = match round2_substitution.get(name) {
+                                None => true,
+                                Some(existing) if existing == ty => {
+                                    // The Round 1 solver already inferred the same
+                                    // value (e.g., from return-context seeding in
+                                    // compute_contextual_types). Mark as return-context
+                                    // so the post-inference retry is suppressed — the
+                                    // callback arguments were already correctly typed.
+                                    had_return_context_substitution = true;
+                                    false
+                                }
+                                Some(existing) => {
+                                    existing == TypeId::UNKNOWN
+                                        || existing == TypeId::ERROR
+                                        || self.inference_type_is_anyish(existing)
+                                        || common::contains_type_parameters(
+                                            self.ctx.types,
+                                            existing,
+                                        )
+                                        || common::contains_infer_types(self.ctx.types, existing)
+                                        || !assign_query::is_fresh_subtype_of(
+                                            self.ctx.types,
+                                            existing,
+                                            ty,
+                                        )
+                                }
+                            };
+
+                            if should_update {
+                                round2_substitution.insert(name, ty);
+                                round2_return_context_names.insert(name);
+                                had_return_context_substitution = true;
+                            }
+                        }
+                    }
+                    for param in &evaluated_shape.params {
+                        for referenced in
+                            common::collect_referenced_types(self.ctx.types, param.type_id)
+                        {
+                            if let Some(info) = common::type_param_info(self.ctx.types, referenced)
+                                && round2_substitution.get(info.name).is_none()
+                            {
+                                let param_name = self.ctx.types.resolve_atom(info.name);
+                                if let Some((_, inferred)) = inferred_type_params_by_name
+                                    .iter()
+                                    .find(|(name, _)| name.as_str() == param_name.as_str())
+                                {
+                                    round2_substitution.insert(info.name, *inferred);
+                                }
+                            }
+                        }
+                    }
+                    trace!("Round 2 substitution prepared");
+
+                    // === Pre-inference from annotated callback parameters ===
+                    // When a callback is context-sensitive (has unannotated params) AND has
+                    // some annotated params, use those annotations to enrich the substitution
+                    // BEFORE computing Round 2 contextual types. This matches tsc's behavior
+                    // where annotated callback params contribute to inference even when the
+                    // callback as a whole is context-sensitive.
+                    //
+                    // Example: test<T extends C>((t1: D, t2) => { t2.test2 })
+                    //   - Round 1 skips the callback (it's sensitive)
+                    //   - But t1: D tells us T = D
+                    //   - Without this, T resolves to constraint C, causing false TS2551
+                    for (i, &arg_idx) in args.iter().enumerate() {
+                        if i < sensitive_args.len()
+                            && sensitive_args[i]
+                            && let Some(shape_param_type) = shape.params.get(i).map(|p| p.type_id)
+                            && let Some(shape_fn) =
+                                query::function_shape_for_type(self.ctx.types, shape_param_type)
+                            && let Some(arg_node) = self.ctx.arena.get(arg_idx)
+                            && let Some(func) = self.ctx.arena.get_function(arg_node)
+                        {
+                            for (j, &param_idx) in func.parameters.nodes.iter().enumerate() {
+                                if let Some(param_node) = self.ctx.arena.get(param_idx)
+                                    && let Some(param) = self.ctx.arena.get_parameter(param_node)
+                                    && param.type_annotation.is_some()
+                                    && let Some(shape_fn_param) = shape_fn.params.get(j)
+                                    && let Some(tp_info) = query::type_parameter_info(
+                                        self.ctx.types,
+                                        shape_fn_param.type_id,
+                                    )
+                                {
+                                    let is_callee_tp =
+                                        shape.type_params.iter().any(|tp| tp.name == tp_info.name);
+                                    // Only override the substitution if it was
+                                    // defaulted to the constraint (not inferred
+                                    // from concrete arguments).
+                                    let existing = substitution.get(tp_info.name);
+                                    let is_defaulted =
+                                        existing.is_none() || existing == tp_info.constraint;
+                                    if is_callee_tp && is_defaulted {
+                                        let ann_type =
+                                            self.get_type_from_type_node(param.type_annotation);
+                                        substitution.insert(tp_info.name, ann_type);
+                                        // Also update round2_substitution so contextual
+                                        // typing of unannotated parameters sees the
+                                        // inferred type from annotations.
+                                        round2_substitution.insert(tp_info.name, ann_type);
+                                        trace!(
+                                            param_index = j,
+                                            ann_type = ann_type.0,
+                                            "Pre-inference: annotated callback param enriched substitution"
+                                        );
+                                    }
+                                }
+                            }
+                        }
+                    }
+
+                    // Sanitize certain function-literal arg types for the second resolve_call
+                    // pass. The sensitive placeholder `(any?) => any` from Round 1 can
+                    // contaminate the solver's type parameter inference when the shape
+                    // param is a bare type parameter or intersection (e.g., `T` or
+                    // `T & Callback`). In those cases, T gets inferred as `(any?) => any`,
+                    // producing Callable types with conflicting call signatures that break
+                    // contextual typing and cause false TS7006 errors.
+                    //
+                    // However, when the shape param is a generic callable like
+                    // `Predicate<A>`, the placeholder's callable structure is useful for
+                    // inferring inner type params (A = any from placeholder params).
+                    // Replacing with UNKNOWN would lose this inference (A = unknown).
+                    //
+                    // Rule: only sanitize when the shape param IS or CONTAINS a top-level
+                    // type parameter (bare T, T & Callable, etc). Leave generic callables
+                    // like Predicate<A> alone since those handle the placeholder correctly.
+                    let sanitized_arg_types: Vec<TypeId> = round1_arg_types
+                        .iter()
+                        .enumerate()
+                        .map(|(i, &ty)| {
+                            if i < sensitive_args.len()
+                                && sensitive_args[i]
+                                && self.is_callback_like_argument(args[i])
+                                && shape.params.get(i).is_some_and(|p| {
+                                    common::is_type_parameter_or_intersection_with_type_parameter(
+                                        self.ctx.types,
+                                        p.type_id,
+                                    )
+                                })
+                            {
+                                TypeId::UNKNOWN
+                            } else {
+                                ty
+                            }
+                        })
+                        .collect();
+                    let round1_instantiated_params = self
+                        .resolve_call_with_checker_adapter(
+                            callee_type_for_context,
+                            &sanitized_arg_types,
+                            force_bivariant_callbacks,
+                            generic_inference_contextual_type,
+                            actual_this_type,
+                        )
+                        .2;
+
+                    if let Some(snap) = &round1_diag_snap {
+                        let round1_end = self.ctx.snapshot_diagnostics();
+                        let preserved_round1_arg_diags =
+                            self.collect_non_callback_diagnostics_between(args, snap, &round1_end);
+                        self.ctx.rollback_diagnostics_filtered(snap, |diag| {
+                            Self::should_preserve_speculative_call_diagnostic(diag)
+                        });
+                        if !preserved_round1_arg_diags.is_empty() {
+                            let mut merged = self.preserved_speculative_call_diagnostics(snap);
+                            self.extend_unique_diagnostics(&mut merged, preserved_round1_arg_diags);
+                            self.ctx.rollback_and_replace_diagnostics(snap, merged);
+                        }
+                    }
+                    if let Some(ts2454_snap) = &round1_ts2454_snap {
+                        self.ctx.restore_ts2454_state(ts2454_snap);
+                    }
+
+                    // === Pre-evaluate instantiated parameter types ===
+                    // After instantiation with Round 1 substitution, parameter types may
+                    // contain unevaluated IndexAccess/KeyOf over Lazy(DefId) references
+                    // (e.g., OptionsForKey[K] → OptionsForKey["a"]). The QueryCache's
+                    // evaluate_type uses NoopResolver which can't resolve Lazy types.
+                    // Use evaluate_type_with_env which resolves Lazy types via the
+                    // TypeEnvironment before evaluation.
+                    let arg_count = args.len();
+                    let has_spread_args = args.iter().any(|&arg_idx| {
+                        self.ctx
+                            .arena
+                            .get(arg_idx)
+                            .is_some_and(|n| n.kind == syntax_kind_ext::SPREAD_ELEMENT)
+                    });
+
+                    let two_pass_result = if !has_spread_args {
+                        let mut progressive_arg_types = round1_arg_types;
+                        let mut round2_arg_types = Vec::with_capacity(arg_count);
+
+                        // Pre-seed the Round 2 contextual substitution from every
+                        // non-context-sensitive argument before contextually typing any
+                        // sensitive callback. The typing loop below otherwise accumulates
+                        // substitutions left-to-right, so a context-sensitive callback
+                        // positioned before the argument that pins a type parameter it
+                        // depends on would be typed with a stale substitution (the parameter
+                        // still uninferred). tsc fixes type parameters from all
+                        // non-context-sensitive arguments before contextually typing any
+                        // context-sensitive argument, making the callback's contextual type
+                        // order-independent.
+                        for (i, &arg_type) in progressive_arg_types.iter().enumerate() {
+                            if sensitive_args.get(i).copied().unwrap_or(false)
+                                || arg_type == TypeId::UNKNOWN
+                                || arg_type == TypeId::ERROR
+                            {
+                                continue;
+                            }
+                            if let Some(shape_param_type) =
+                                shape.params.get(i).map(|p| p.type_id).or_else(|| {
+                                    let last = shape.params.last()?;
+                                    last.rest.then_some(last.type_id)
+                                })
+                            {
+                                self.merge_arg_return_context_into_round2(
+                                    &shape,
+                                    shape_param_type,
+                                    arg_type,
+                                    &tracked_type_params,
+                                    false,
+                                    &mut round2_substitution,
+                                );
+                            }
+                        }
+
+                        for (i, &arg_idx) in args.iter().enumerate() {
+                            if sensitive_args.get(i).copied().unwrap_or(false)
+                                && let Some(first_branch_idx) =
+                                    self.callback_first_conditional_branch(arg_idx)
+                                && let Some(param_type) = shape.params.get(i).map(|p| p.type_id)
+                                && let Some(callback_shape) =
+                                    query::function_shape_for_type(self.ctx.types, param_type)
+                            {
+                                let first_branch_type = self.get_type_of_node(first_branch_idx);
+                                let mut first_branch_substitution =
+                                    crate::query_boundaries::common::TypeSubstitution::new();
+                                let mut visited = FxHashSet::default();
+                                self.collect_return_context_substitution(
+                                    callback_shape.return_type,
+                                    first_branch_type,
+                                    &tracked_type_params,
+                                    &mut first_branch_substitution,
+                                    &mut visited,
+                                );
+                                for (&name, &ty) in first_branch_substitution.map().iter() {
+                                    let should_update = match round2_substitution.get(name) {
+                                        None => true,
+                                        Some(existing) if existing == ty => false,
+                                        Some(existing) => {
+                                            existing == TypeId::UNKNOWN
+                                                || common::contains_type_parameters(
+                                                    self.ctx.types,
+                                                    existing,
+                                                )
+                                                || common::contains_infer_types(
+                                                    self.ctx.types,
+                                                    existing,
+                                                )
+                                        }
+                                    };
+                                    if ty != TypeId::UNKNOWN
+                                        && ty != TypeId::ERROR
+                                        && !common::contains_type_parameters(self.ctx.types, ty)
+                                        && !common::contains_infer_types(self.ctx.types, ty)
+                                        && should_update
+                                    {
+                                        round2_substitution.insert(name, ty);
+                                    }
+                                }
+                            }
+                            if sensitive_args.get(i).copied().unwrap_or(false) {
+                                self.clear_contextual_resolution_cache();
+                                self.invalidate_expression_for_contextual_retry(arg_idx);
+                            }
+                            let mut contextual_substitution = self
+                                .widen_round2_contextual_substitution(&shape, &round2_substitution);
+                            self.restore_conflicting_direct_literal_substitutions(
+                                &mut contextual_substitution,
+                                &direct_literal_conflict_type_params,
+                            );
+                            if !round2_return_context_names.is_empty()
+                                && let Some(param_type) =
+                                    shape.params.get(i).map(|p| p.type_id).or_else(|| {
+                                        let last = shape.params.last()?;
+                                        last.rest.then_some(last.type_id)
+                                    })
+                            {
+                                let names_to_strip: FxHashSet<_> = self
+                                    .sensitive_callback_nested_parameter_type_params(
+                                        &shape, param_type,
+                                    )
+                                    .into_iter()
+                                    .filter(|name| round2_return_context_names.contains(name))
+                                    .collect();
+                                if !names_to_strip.is_empty() {
+                                    let mut filtered =
+                                        crate::query_boundaries::common::TypeSubstitution::new();
+                                    for (&name, &type_id) in contextual_substitution.map() {
+                                        if !names_to_strip.contains(&name) {
+                                            filtered.insert(name, type_id);
+                                        }
+                                    }
+                                    contextual_substitution = filtered;
+                                }
+                            }
+                            let round2_contextual_types = self.compute_round2_contextual_types(
+                                &shape,
+                                round1_instantiated_params.as_deref(),
+                                &sensitive_args,
+                                &contextual_substitution,
+                                arg_count,
+                            );
+                            let expected_type = round2_contextual_types
+                                .get(i)
+                                .copied()
+                                .flatten()
+                                .or_else(|| base_contextual_param_types.get(i).copied().flatten());
+                            let arg_type = if expected_type.is_some_and(|expected| {
+                                expected != TypeId::UNKNOWN
+                                    && expected != TypeId::ERROR
+                                    && !common::contains_infer_types(self.ctx.types, expected)
+                                    && !common::contains_type_parameters(self.ctx.types, expected)
+                            }) {
+                                let (start, end) = self
+                                    .ctx
+                                    .arena
+                                    .get(arg_idx)
+                                    .map(|node| (node.pos, node.end))
+                                    .unwrap_or((0, 0));
+                                self.ctx.diagnostics.retain(|diag| {
+                                    diag.start < start
+                                        || diag.start >= end
+                                        // TS2454 (variable used before being assigned) is a
+                                        // semantic fact about the variable, not a speculative
+                                        // inference artifact. Preserve it across round 2
+                                        // re-typing so it isn't silently dropped.
+                                        || diag.code
+                                            == diagnostic_codes::VARIABLE_IS_USED_BEFORE_BEING_ASSIGNED
+                                        // TS2872/TS2873 (always truthy/falsy) are purely
+                                        // syntactic facts, not speculative results.
+                                        || diag.code
+                                            == diagnostic_codes::THIS_KIND_OF_EXPRESSION_IS_ALWAYS_TRUTHY
+                                        || diag.code
+                                            == diagnostic_codes::THIS_KIND_OF_EXPRESSION_IS_ALWAYS_FALSY
+                                        // TS7006/TS7019 (parameter implicitly has 'any' type) is
+                                        // a semantic fact about the parameter when the contextual
+                                        // callable signature has fewer parameters than the callback.
+                                        // Preserve these since re-typing won't change the fact.
+                                        || diag.code
+                                            == diagnostic_codes::PARAMETER_IMPLICITLY_HAS_AN_TYPE
+                                        || diag.code
+                                            == diagnostic_codes::REST_PARAMETER_IMPLICITLY_HAS_AN_ANY_TYPE
+                                });
+                                self.ctx.rebuild_emitted_diagnostics_from_current();
+                                self.compute_single_call_argument_type(
+                                    arg_idx,
+                                    expected_type,
+                                    check_excess_properties,
+                                    i,
+                                    args.len(),
+                                    true,
+                                    callable_ctx,
+                                )
+                            } else {
+                                self.compute_single_call_argument_type(
+                                    arg_idx,
+                                    expected_type,
+                                    check_excess_properties,
+                                    i,
+                                    args.len(),
+                                    true,
+                                    callable_ctx,
+                                )
+                            };
+                            let preserve_round1_object_literal =
+                                self.ctx.arena.get(arg_idx).is_some_and(|node| {
+                                    node.kind == syntax_kind_ext::OBJECT_LITERAL_EXPRESSION
+                                        && !is_contextually_sensitive(self, arg_idx)
+                                });
+                            let arg_type_for_refinement = if preserve_round1_object_literal {
+                                progressive_arg_types.get(i).copied().unwrap_or(arg_type)
+                            } else {
+                                expected_type
+                                    .map(|expected| {
+                                        if self
+                                            .target_has_concrete_return_context_for_generic_refinement(
+                                                expected,
+                                            )
+                                        {
+                                            self.instantiate_generic_function_argument_against_target_for_refinement(
+                                                arg_type, expected,
+                                            )
+                                        } else {
+                                            self.instantiate_generic_function_argument_against_target_params(
+                                                arg_type, expected,
+                                            )
+                                        }
+                                    })
+                                    .unwrap_or(arg_type)
+                            };
+                            trace!(
+                                arg_index = i,
+                                expected_type = ?expected_type.map(|t| t.0),
+                                expected_type_display = ?expected_type.map(|t| self.format_type(t)),
+                                arg_type = arg_type.0,
+                                arg_type_display = %self.format_type(arg_type),
+                                "Round 2: recomputed argument type"
+                            );
+                            round2_arg_types.push(arg_type_for_refinement);
+                            if i < progressive_arg_types.len() {
+                                progressive_arg_types[i] = arg_type_for_refinement;
+                            }
+
+                            if let Some(shape_param_type) =
+                                shape.params.get(i).map(|p| p.type_id).or_else(|| {
+                                    let last = shape.params.last()?;
+                                    last.rest.then_some(last.type_id)
+                                })
+                            {
+                                self.merge_arg_return_context_into_round2(
+                                    &shape,
+                                    shape_param_type,
+                                    arg_type_for_refinement,
+                                    &tracked_type_params,
+                                    sensitive_args.get(i).copied().unwrap_or(false),
+                                    &mut round2_substitution,
+                                );
+                            }
+
+                            let expected_still_unresolved = expected_type.is_some_and(|expected| {
+                                common::contains_infer_types(self.ctx.types, expected)
+                                    || common::contains_type_parameters(self.ctx.types, expected)
+                            });
+                            let arg_is_callable =
+                                common::is_callable_type(self.ctx.types, arg_type);
+                            let skip_return_only_refinement = self
+                                .ctx
+                                .arena
+                                .get(arg_idx)
+                                .and_then(|node| self.ctx.arena.get_function(node))
+                                .is_some_and(|func| {
+                                    func.parameters.nodes.is_empty()
+                                        && func.type_annotation.is_none()
+                                });
+                            let should_refine_substitution =
+                                sensitive_args.get(i).copied().unwrap_or(false)
+                                    || (expected_still_unresolved && arg_is_callable);
+                            if should_refine_substitution && !skip_return_only_refinement {
+                                let refined_substitution = {
+                                    let env = self.ctx.type_env.borrow();
+                                    call_checker::compute_contextual_types_with_context(
+                                        self.ctx.types,
+                                        &self.ctx,
+                                        &env,
+                                        &evaluated_shape,
+                                        &progressive_arg_types,
+                                        generic_inference_contextual_type,
+                                    )
+                                };
+                                let mut substitution_changed = false;
+                                for (&name, &ty) in refined_substitution.map().iter() {
+                                    if ty == TypeId::UNKNOWN
+                                        || (sensitive_args.get(i).copied().unwrap_or(false)
+                                            && (ty == TypeId::NULL || ty == TypeId::UNDEFINED))
+                                        || common::contains_infer_types(self.ctx.types, ty)
+                                        || common::contains_type_parameters(self.ctx.types, ty)
+                                    {
+                                        continue;
+                                    }
+
+                                    let should_update = match round2_substitution.get(name) {
+                                        None => true,
+                                        Some(existing) if existing == ty => false,
+                                        Some(existing) => {
+                                            existing == TypeId::UNKNOWN
+                                                || common::contains_infer_types(
+                                                    self.ctx.types,
+                                                    existing,
+                                                )
+                                                || common::contains_type_parameters(
+                                                    self.ctx.types,
+                                                    existing,
+                                                )
+                                        }
+                                    };
+
+                                    if should_update {
+                                        round2_substitution.insert(name, ty);
+                                        substitution_changed = true;
+                                    }
+                                }
+                                if substitution_changed {
+                                    trace!("Round 2 substitution refined");
+                                }
+                            }
+                        }
+
+                        round2_arg_types
+                    } else {
+                        let mut contextual_substitution =
+                            self.widen_round2_contextual_substitution(&shape, &round2_substitution);
+                        self.restore_conflicting_direct_literal_substitutions(
+                            &mut contextual_substitution,
+                            &direct_literal_conflict_type_params,
+                        );
+                        let round2_contextual_types = self.compute_round2_contextual_types(
+                            &shape,
+                            round1_instantiated_params.as_deref(),
+                            &sensitive_args,
+                            &contextual_substitution,
+                            arg_count,
+                        );
+
+                        self.collect_call_argument_types_with_context(
+                            args,
+                            |i, _arg_count| {
+                                if i < round2_contextual_types.len() {
+                                    round2_contextual_types[i]
+                                } else {
+                                    base_contextual_param_types.get(i).copied().flatten()
+                                }
+                            },
+                            check_excess_properties,
+                            None,
+                            callable_ctx,
+                        )
+                    };
+                    // Snapshot the checker's intra-expression Round 1 substitution
+                    // so the post-solver path can refine `instantiated_params` when
+                    // the solver's single-pass inference loses bindings the checker
+                    // established (e.g., `O = setup-derived-shape` is dropped
+                    // because `Unwrap<O>` in `map`'s return position fails to
+                    // reverse-infer).
+                    //
+                    // Gate on `extracted_round1_partials.iter().any()` — only when
+                    // at least one argument contributed a Round 1 partial type
+                    // (a non-sensitive object-literal property fed inference) is
+                    // there reason to believe the checker has information the
+                    // solver lacks. Otherwise (e.g., a `Promise.then(() => x, () => 1)`
+                    // call with two sensitive zero-param callbacks), Round 1's
+                    // substitution is derived from the same arg types the solver
+                    // sees at its boundary; overriding would clobber correct
+                    // solver inferences.
+                    let any_round1_partial = extracted_round1_partials.iter().any(|&b| b);
+                    if any_round1_partial && !substitution.is_empty() {
+                        checker_round2_substitution = Some(substitution.clone());
+                        checker_round2_shape = Some(shape.clone());
+                    }
+                    two_pass_result
+                } else {
+                    // Single-pass generic calls still erase type params from empty-array
+                    // contextual types so `[]` does not feed raw `T[]` back into inference.
+                    let type_param_eraser = {
+                        use crate::query_boundaries::common::TypeSubstitution;
+                        let mut sub = TypeSubstitution::new();
+                        for tp in &shape.type_params {
+                            sub.insert(tp.name, tp.constraint.unwrap_or(TypeId::UNKNOWN));
+                        }
+                        sub
+                    };
+                    let arena = self.ctx.arena;
+                    let single_pass_contextual_types: Vec<Option<TypeId>> = (0..args.len())
+                        .map(|i| {
+                            let param_type =
+                                base_contextual_param_types.get(i).copied().flatten()?;
+                            let is_empty_array_literal = arena.get(args[i]).is_some_and(|n| {
+                                n.kind == syntax_kind_ext::ARRAY_LITERAL_EXPRESSION
+                                    && arena
+                                        .get_literal_expr(n)
+                                        .is_some_and(|lit| lit.elements.nodes.is_empty())
+                            });
+                            let param_type = if is_empty_array_literal {
+                                use crate::query_boundaries::common::instantiate_type;
+                                instantiate_type(self.ctx.types, param_type, &type_param_eraser)
+                            } else {
+                                param_type
+                            };
+                            Some(self.normalize_contextual_call_param_type(param_type))
+                        })
+                        .collect();
+                    let single_pass_return_context_substitution =
+                        if contextual_type.is_some_and(|ctx| {
+                            let contextual_generic_callable =
+                                call_checker::get_contextual_signature(self.ctx.types, ctx)
+                                    .is_some_and(|shape| !shape.type_params.is_empty());
+                            (!common::contains_type_parameters(self.ctx.types, ctx)
+                                || contextual_generic_callable)
+                                && !common::contains_infer_types(self.ctx.types, ctx)
+                                && !common::contains_type_by_id(
+                                    self.ctx.types,
+                                    ctx,
+                                    TypeId::UNKNOWN,
+                                )
+                        }) {
+                            self.compute_return_context_substitution_from_shape(
+                                &shape,
+                                contextual_type,
+                            )
+                        } else {
+                            crate::query_boundaries::common::TypeSubstitution::new()
+                        };
+                    let needs_refresh = args.iter().enumerate().any(|(i, &arg)| {
+                        self.argument_needs_refresh_for_contextual_call(
+                            arg,
+                            base_contextual_param_types.get(i).copied().flatten(),
+                        )
+                    }) || !single_pass_return_context_substitution.is_empty();
+                    let initial_arg_snap = needs_refresh.then(|| self.ctx.snapshot_diagnostics());
+                    let initial_ts2454_snap = initial_arg_snap
+                        .as_ref()
+                        .map(|_| self.ctx.emitted_ts2454_errors.clone());
+                    let preserved_object_literal_implicit_any_spans: Vec<_> = args
+                        .iter()
+                        .copied()
+                        .filter(|&arg_idx| {
+                            self.ctx.arena.get(arg_idx).is_some_and(|node| {
+                                node.kind == syntax_kind_ext::OBJECT_LITERAL_EXPRESSION
+                            })
+                        })
+                        .flat_map(|arg_idx| {
+                            self.object_literal_noncontextual_function_param_spans(arg_idx)
+                        })
+                        .collect();
+                    let mut initial_arg_types = self.collect_call_argument_types_with_context(
+                        args,
+                        |i, _arg_count| {
+                            if i < single_pass_contextual_types.len() {
+                                single_pass_contextual_types[i]
+                            } else {
+                                base_contextual_param_types.get(i).copied().flatten()
+                            }
+                        },
+                        check_excess_properties,
+                        None, // No skipping needed for single-pass
+                        callable_ctx,
+                    );
+                    for (i, arg_type) in initial_arg_types.iter_mut().enumerate() {
+                        let Some(&arg_idx) = args.get(i) else {
+                            continue;
+                        };
+                        let Some(arg_node) = self.ctx.arena.get(arg_idx) else {
+                            continue;
+                        };
+                        if !self.is_callback_like_argument(arg_idx) {
+                            continue;
+                        }
+                        if self
+                            .ctx
+                            .arena
+                            .get_function(arg_node)
+                            .and_then(|func| func.type_parameters.as_ref())
+                            .is_some_and(|params| !params.nodes.is_empty())
+                        {
+                            let raw_arg_type =
+                                self.get_type_of_node_with_request(arg_idx, &TypingRequest::NONE);
+                            let seeded =
+                                self.sanitize_generic_inference_arg_type(arg_idx, raw_arg_type);
+                            if seeded != TypeId::UNKNOWN && seeded != TypeId::ERROR {
+                                *arg_type = seeded;
+                            }
+                        }
+                    }
+                    if !needs_refresh {
+                        initial_arg_types
+                    } else {
+                        let refreshed_direct_function_arg_spans: Vec<_> = args
+                            .iter()
+                            .enumerate()
+                            .filter_map(|(i, &arg_idx)| {
+                                self.argument_needs_refresh_for_contextual_call(
+                                    arg_idx,
+                                    base_contextual_param_types.get(i).copied().flatten(),
+                                )
+                                .then(|| {
+                                    self.ctx.arena.get(arg_idx).and_then(|node| {
+                                        (node.kind == syntax_kind_ext::ARROW_FUNCTION
+                                            || node.kind == syntax_kind_ext::FUNCTION_EXPRESSION)
+                                            .then_some((node.pos, node.end))
+                                    })
+                                })
+                                .flatten()
+                            })
+                            .collect();
+                        let return_context_substitution = single_pass_return_context_substitution;
+                        if !return_context_substitution.is_empty() {
+                            if let Some(snap) = &initial_arg_snap {
+                                let initial_arg_end = self.ctx.snapshot_diagnostics();
+                                let preserved_initial_arg_diags = self
+                                    .collect_non_callback_diagnostics_between(
+                                        args,
+                                        snap,
+                                        &initial_arg_end,
+                                    );
+                                self.ctx.rollback_diagnostics_filtered(snap, |diag| {
+                                    Self::should_preserve_speculative_call_diagnostic(diag)
+                                        || matches!(
+                                            diag.code,
+                                            diagnostic_codes::PARAMETER_IMPLICITLY_HAS_AN_TYPE
+                                                | diagnostic_codes::REST_PARAMETER_IMPLICITLY_HAS_AN_ANY_TYPE
+                                                | diagnostic_codes::BINDING_ELEMENT_IMPLICITLY_HAS_AN_TYPE
+                                                | diagnostic_codes::PARAMETER_HAS_A_NAME_BUT_NO_TYPE_DID_YOU_MEAN
+                                        ) && preserved_object_literal_implicit_any_spans.iter().any(
+                                            |(start, end)| diag.start >= *start && diag.start < *end,
+                                        ) || matches!(
+                                            diag.code,
+                                            diagnostic_codes::OBJECT_LITERAL_MAY_ONLY_SPECIFY_KNOWN_PROPERTIES_AND_DOES_NOT_EXIST_IN_TYPE
+                                                | diagnostic_codes::OBJECT_LITERAL_MAY_ONLY_SPECIFY_KNOWN_PROPERTIES_BUT_DOES_NOT_EXIST_IN_TYPE_DID
+                                        ) && !refreshed_direct_function_arg_spans
+                                            .iter()
+                                            .any(|(start, end)| diag.start >= *start && diag.start < *end)
+                                });
+                                if !preserved_initial_arg_diags.is_empty() {
+                                    let mut merged =
+                                        self.preserved_speculative_call_diagnostics(snap);
+                                    self.extend_unique_diagnostics(
+                                        &mut merged,
+                                        preserved_initial_arg_diags,
+                                    );
+                                    self.ctx.rollback_and_replace_diagnostics(snap, merged);
+                                }
+                            }
+                            if let Some(ts2454_snap) = &initial_ts2454_snap {
+                                self.ctx.restore_ts2454_state(ts2454_snap);
+                            }
+                            self.clear_contextual_resolution_cache();
+                            for (i, &arg_idx) in args.iter().enumerate() {
+                                if self.argument_needs_refresh_for_contextual_call(
+                                    arg_idx,
+                                    base_contextual_param_types.get(i).copied().flatten(),
+                                ) {
+                                    self.invalidate_expression_for_contextual_retry(arg_idx);
+                                }
+                            }
+                            let refreshed_contextual_types: Vec<Option<TypeId>> = (0..args.len())
+                                .map(|i| {
+                                    let param =
+                                        shape.params.get(i).map(|p| (p.type_id, p.rest)).or_else(
+                                            || {
+                                                let last = shape.params.last()?;
+                                                last.rest.then_some((last.type_id, true))
+                                            },
+                                        )?;
+                                    let instantiated =
+                                        crate::query_boundaries::common::instantiate_type(
+                                            self.ctx.types,
+                                            param.0,
+                                            &return_context_substitution,
+                                        );
+                                    let param_type = if param.1 {
+                                        self.rest_argument_element_type_with_env(instantiated)
+                                    } else {
+                                        instantiated
+                                    };
+                                    Some(self.normalize_contextual_call_param_type(param_type))
+                                })
+                                .collect();
+                            let mut refreshed_args = self.collect_call_argument_types_with_context(
+                                args,
+                                |i, _arg_count| {
+                                    refreshed_contextual_types
+                                        .get(i)
+                                        .copied()
+                                        .flatten()
+                                        .or_else(|| {
+                                            base_contextual_param_types.get(i).copied().flatten()
+                                        })
+                                },
+                                check_excess_properties,
+                                None,
+                                callable_ctx,
+                            );
+                            // When the return context substitution provides concrete
+                            // types for type parameters, adjust callback arg types
+                            // whose return type is a fresh literal that satisfies the
+                            // contextual return type. This matches tsc's behavior
+                            // where literals contextually typed by type parameters
+                            // are treated as non-fresh during inference, preventing
+                            // widening (e.g., "ELSE" → string).
+                            for (i, arg_type) in refreshed_args.iter_mut().enumerate() {
+                                let Some(&arg_idx) = args.get(i) else {
+                                    continue;
+                                };
+                                let Some(_) = self.ctx.arena.get(arg_idx) else {
+                                    continue;
+                                };
+                                if !self.is_callback_like_argument(arg_idx) {
+                                    continue;
+                                }
+                                let Some(fn_shape) =
+                                    common::function_shape_for_type(self.ctx.types, *arg_type)
+                                else {
+                                    continue;
+                                };
+                                // Only adjust when the return type contains
+                                // literals that the solver would widen.
+                                let ret = fn_shape.return_type;
+                                let widened = tsz_solver::operations::widening::widen_literal_type(
+                                    self.ctx.types,
+                                    ret,
+                                );
+                                if widened == ret {
+                                    // No widening would happen, skip
+                                    continue;
+                                }
+                                // Get the contextual return type for this callback
+                                // from the refreshed param type
+                                let ctx_return = refreshed_contextual_types
+                                    .get(i)
+                                    .copied()
+                                    .flatten()
+                                    .and_then(|ctx| {
+                                        crate::query_boundaries::checkers::call::get_contextual_signature(self.ctx.types, ctx)
+                                    })
+                                    .map(|s| s.return_type);
+                                let Some(ctx_return) = ctx_return else {
+                                    continue;
+                                };
+                                // If the literal satisfies the contextual return type,
+                                // replace the callback return with the contextual type
+                                // to prevent widening in the solver's inference.
+                                if self.is_assignable_to_with_env(fn_shape.return_type, ctx_return)
+                                {
+                                    let mut new_shape = (*fn_shape).clone();
+                                    new_shape.return_type = ctx_return;
+                                    *arg_type = self.ctx.types.factory().function(new_shape);
+                                }
+                            }
+                            refreshed_args
+                        } else if let Some(instantiated_params) = self
+                            .resolve_call_with_checker_adapter(
+                                callee_type_for_context,
+                                &initial_arg_types,
+                                force_bivariant_callbacks,
+                                contextual_type,
+                                actual_this_type,
+                            )
+                            .2
+                        {
+                            if let Some(snap) = &initial_arg_snap {
+                                let initial_arg_end = self.ctx.snapshot_diagnostics();
+                                let preserved_initial_arg_diags = self
+                                    .collect_non_callback_diagnostics_between(
+                                        args,
+                                        snap,
+                                        &initial_arg_end,
+                                    );
+                                self.ctx.rollback_diagnostics_filtered(snap, |diag| {
+                                    Self::should_preserve_speculative_call_diagnostic(diag)
+                                        || matches!(
+                                            diag.code,
+                                            diagnostic_codes::PARAMETER_IMPLICITLY_HAS_AN_TYPE
+                                                | diagnostic_codes::REST_PARAMETER_IMPLICITLY_HAS_AN_ANY_TYPE
+                                                | diagnostic_codes::BINDING_ELEMENT_IMPLICITLY_HAS_AN_TYPE
+                                                | diagnostic_codes::PARAMETER_HAS_A_NAME_BUT_NO_TYPE_DID_YOU_MEAN
+                                        ) && preserved_object_literal_implicit_any_spans.iter().any(
+                                            |(start, end)| diag.start >= *start && diag.start < *end,
+                                        ) || matches!(
+                                            diag.code,
+                                            diagnostic_codes::OBJECT_LITERAL_MAY_ONLY_SPECIFY_KNOWN_PROPERTIES_AND_DOES_NOT_EXIST_IN_TYPE
+                                                | diagnostic_codes::OBJECT_LITERAL_MAY_ONLY_SPECIFY_KNOWN_PROPERTIES_BUT_DOES_NOT_EXIST_IN_TYPE_DID
+                                        ) && !refreshed_direct_function_arg_spans
+                                            .iter()
+                                            .any(|(start, end)| diag.start >= *start && diag.start < *end)
+                                });
+                                if !preserved_initial_arg_diags.is_empty() {
+                                    let mut merged =
+                                        self.preserved_speculative_call_diagnostics(snap);
+                                    self.extend_unique_diagnostics(
+                                        &mut merged,
+                                        preserved_initial_arg_diags,
+                                    );
+                                    self.ctx.rollback_and_replace_diagnostics(snap, merged);
+                                }
+                            }
+                            if let Some(ts2454_snap) = &initial_ts2454_snap {
+                                self.ctx.restore_ts2454_state(ts2454_snap);
+                            }
+                            self.clear_contextual_resolution_cache();
+                            for (i, &arg_idx) in args.iter().enumerate() {
+                                if self.argument_needs_refresh_for_contextual_call(
+                                    arg_idx,
+                                    base_contextual_param_types.get(i).copied().flatten(),
+                                ) {
+                                    self.invalidate_expression_for_contextual_retry(arg_idx);
+                                }
+                            }
+                            let instantiated_params = self
+                                .resolve_signature_parameter_type_queries(
+                                    &shape.params,
+                                    &instantiated_params,
+                                );
+                            let refreshed_contextual_types = self
+                                .contextual_param_types_from_instantiated_params(
+                                    &instantiated_params,
+                                    args.len(),
+                                )
+                                .into_iter()
+                                .map(|param_type| {
+                                    param_type.map(|param_type| {
+                                        self.normalize_contextual_call_param_type(param_type)
+                                    })
+                                })
+                                .collect::<Vec<_>>();
+                            let refreshed_arg_types = self
+                                .collect_call_argument_types_with_context(
+                                    args,
+                                    |i, _arg_count| {
+                                        refreshed_contextual_types
+                                            .get(i)
+                                            .copied()
+                                            .flatten()
+                                            .or_else(|| {
+                                                base_contextual_param_types
+                                                    .get(i)
+                                                    .copied()
+                                                    .flatten()
+                                            })
+                                    },
+                                    check_excess_properties,
+                                    None,
+                                    callable_ctx,
+                                );
+                            self.refine_generic_function_args_against_instantiated_params(
+                                refreshed_arg_types,
+                                &instantiated_params,
+                            )
+                        } else {
+                            initial_arg_types
+                        }
+                    }
+                }
+            } else {
+                // Shouldn't happen for generic call detection, but keep single-pass fallback.
+                let single_pass_contextual_types: Vec<Option<TypeId>> = (0..args.len())
+                    .map(|i| {
+                        let param_type = base_contextual_param_types.get(i).copied().flatten()?;
+                        Some(self.normalize_contextual_call_param_type(param_type))
+                    })
+                    .collect();
+                self.collect_call_argument_types_with_context(
+                    args,
+                    |i, _arg_count| {
+                        if i < single_pass_contextual_types.len() {
+                            single_pass_contextual_types[i]
+                        } else {
+                            base_contextual_param_types.get(i).copied().flatten()
+                        }
+                    },
+                    check_excess_properties,
+                    None, // No skipping needed for single-pass
+                    callable_ctx,
+                )
+            }
+        } else {
+            // === Single-pass: Standard argument collection ===
+            // Non-generic calls or calls with explicit type arguments use the standard flow.
+            let single_pass_contextual_types: Vec<Option<TypeId>> = (0..args.len())
+                .map(|i| {
+                    let param_type = base_contextual_param_types.get(i).copied().flatten()?;
+                    Some(self.normalize_contextual_call_param_type(param_type))
+                })
+                .collect();
+            let nominal_lib_object_callback_context: Vec<bool> = single_pass_contextual_types
+                .iter()
+                .map(|param_type| {
+                    param_type.is_some_and(|ty| {
+                        self.nominal_lib_object_callback_return_type(ty).is_some()
+                    })
+                })
+                .collect();
+            non_generic_contextual_types = Some(single_pass_contextual_types.clone());
+            self.collect_call_argument_types_with_context(
+                args,
+                |i, _arg_count| {
+                    if nominal_lib_object_callback_context
+                        .get(i)
+                        .copied()
+                        .unwrap_or(false)
+                    {
+                        return None;
+                    }
+                    if i < single_pass_contextual_types.len() {
+                        single_pass_contextual_types[i]
+                    } else {
+                        base_contextual_param_types.get(i).copied().flatten()
+                    }
+                },
+                check_excess_properties,
+                None, // No skipping needed for single-pass
+                callable_ctx,
+            )
+        };
+        (
+            non_generic_contextual_types,
+            pushed_this_type_from_shape,
+            had_return_context_substitution,
+            checker_round2_substitution,
+            checker_round2_shape,
+            direct_literal_conflict_substitution,
+            arg_types,
+        )
+}
